Only show "Delete Cache" link to MU site admins or admins on WP blogs
git-svn-id: http://plugins.svn.wordpress.org/wp-super-cache/trunk@66448 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wp-cache.php
@@ -1076,6 +1076,12 @@ function wp_cache_files() {
 }
 
 function delete_cache_dashboard() {
+	if( function_exists( 'is_site_admin' ) && !is_site_admin() )
+		return false;
+
+	if( function_exists('current_user_can') && !current_user_can('manage_options') )
+		return false;
+
 	echo "<li><a href='" . wp_nonce_url( 'options-general.php?page=wpsupercache&wp_delete_cache=1#list', 'wp-cache' ) . "' target='_blank' title='Delete Super Cache cached files (opens in new window)'>Delete Cache</a></li>";
 }
 add_action( 'dashmenu', 'delete_cache_dashboard' );
